duui-annis-reader/requirements.txt
# graphannis
pandas
fastapi
pydantic
pydantic-settings
starlette
uvicorn
dkpro-cassis
python-multipart
websockets
tqdm
Deprecated